nit: declare vInt as the desired type

+++ codec/packer_test.go
@@ -122,16 +122,16 @@ func TestPackerAddress(t *testing.T) {
 
 func TestNewReader(t *testing.T) {
 	require := require.New(t)
-	vInt := 900
+	vInt := int32(900)
 	wp := NewWriter(5, 5)
 	// Add an int and a bool
-	wp.PackInt(int32(vInt))
+	wp.PackInt(vInt)
 	wp.PackBool(true)
 	// Create reader
 	rp := NewReader(wp.Bytes(), 2)
 	require.Equal(wp.Bytes(), rp.Bytes(), "Reader not initialized correctly.")
 	// Unpack both values
-	require.Equal(int32(vInt), rp.UnpackInt(true), "Reader unpacked correctly.")
+	require.Equal(vInt, rp.UnpackInt(true), "Reader unpacked correctly.")
 	require.True(rp.UnpackBool(), "Reader unpacked correctly.")
 	require.NoError(rp.Err(), "Reader set error during unpack.")
 	// Unpacked not packed with required
